refactor: Implement code quality improvements
This commit addresses various code quality improvements across the repository:

- Refactored duplicated code by introducing shared modules and functions.
- Updated ESLint configuration to enforce stricter rules and fixed all linting errors.
- Added or updated JSDoc comments for better code documentation.
- Removed non-JSDoc inline comments.
- Updated `package.json` with new scripts for code analysis (duplicates, unused dependencies, coverage) and incremented the application version to 2.0.2.
- Updated `AGENTS.md` and `constitution.md` to reflect new development guidelines and quality gates.
- Marked all tasks in `tasks.md` and progress in `plan.md` as complete.
