+++ .gitignore
@@ -2,4 +2,4 @@ node_modules
 .DS_Store
 dist
 .serena
-
+coverage/
