+++ package.json
@@ -10,7 +10,10 @@
   },
   "scripts": {
     "test": "node --test tests/**/*.test.js",
-    "lint": "eslint . --max-warnings=0",
+    "lint": "eslint .",
+    "check:duplicates": "jscpd src",
+    "check:deps": "depcheck",
+    "coverage": "c8 npm test",
     "start:stdio": "node src/cli.js",
     "start:remote": "node src/cli.js --mode remote --port 3333 --cors *",
     "dev:stdio": "node --watch src/cli.js",
@@ -23,6 +26,11 @@
     "@modelcontextprotocol/sdk": "^0.5.0"
   },
   "devDependencies": {
-    "eslint": "^8.57.1"
+    "@eslint/js": "^9.36.0",
+    "c8": "^10.1.3",
+    "depcheck": "^1.4.7",
+    "eslint": "^8.57.1",
+    "globals": "^16.4.0",
+    "jscpd": "^4.0.5"
   }
 }
