+++ specs/004-code-quality-improvements/plan.md
@@ -1,4 +1,3 @@
-
 # Implementation Plan: Code Quality Improvements
 
 **Branch**: `004-code-quality-improvements` | **Date**: 2025-09-22 | **Spec**: [link](./spec.md)
@@ -35,7 +34,7 @@ The feature focuses on a comprehensive review and refactoring of the existing co
 **Storage**: N/A
 **Testing**: `node --test`
 **Target Platform**: Node.js CLI
-**Project Type**: Single project
+**Project Type**: Single project.
 **Performance Goals**: N/A
 **Constraints**: Must not alter existing behavior as validated by current tests.
 **Scale/Scope**: Entire repository codebase.
@@ -120,13 +119,13 @@ No constitutional violations that require justification.
 **Phase Status**:
 - [X] Phase 0: Research complete (/plan command)
 - [X] Phase 1: Design complete (/plan command)
-- [ ] Phase 2: Task planning complete (/plan command - describe approach only)
-- [ ] Phase 3: Tasks generated (/tasks command)
-- [ ] Phase 4: Implementation complete
-- [ ] Phase 5: Validation passed
+- [X] Phase 2: Task planning complete (/plan command - describe approach only)
+- [X] Phase 3: Tasks generated (/tasks command)
+- [X] Phase 4: Implementation complete
+- [X] Phase 5: Validation passed
 
 **Gate Status**:
 - [X] Initial Constitution Check: PASS
-- [ ] Post-Design Constitution Check: PASS
+- [X] Post-Design Constitution Check: PASS
 - [X] All NEEDS CLARIFICATION resolved
-- [ ] Complexity deviations documented
+- [X] Complexity deviations documented
